+++ backend/common/pkg/models/paper_ownership.go
@@ -1,9 +1,9 @@
 package models
 
 type PaperOwnership struct {
-	ID      int `gorm:"primaryKey"`
-	UserID  int
-	PaperID int
+	ID      int64 `gorm:"primaryKey"`
+	UserID  int64
+	PaperID int64
 	Path    string
 	Type    string `gorm:"type:varchar(10);not null;check:type IN ('OWNER', 'SHARED')"`
 }
