+++ backend/server/internal/dtos/question_category.dto.go
@@ -5,10 +5,10 @@ type UpdateCategoryDto struct {
 }
 
 type ReorderCategoryDto struct {
-	Categories []int `json:"categories" validate:"required,min=1"`
+	Categories []int64 `json:"categories" validate:"required,min=1"`
 }
 
 type CategoryOrderDto struct {
-	ID    int `json:"id" validate:"required"`
-	Order int `json:"order" validate:"required"`
+	ID    int64 `json:"id" validate:"required"`
+	Order int   `json:"order" validate:"required"`
 }
